Add files via upload
SONAR Rock v/s Mine Prediction, Project description

consider there is a submarine of a country trying to enter into ocean borders of some other country. And enemy country has planted some ocean mines which would explode if any object comes into its contact. there also can be rocks in the ocean so, submarine needs to predict whether it is crossing Rock or Mine.

Submarine uses Sonar that sends sound signals, these signals are than processed to detect whether the object is a Mine or Rock in the Ocean.

job of the model is to predict the object beneath the ocean is Rock or Mine.

DATA -> Data is collected from Kaggle. Mines are made up of metal. Sonar data which is obtained from Rock and a metal cylinder will be feeded to machine learning model which will be used to predict the object is Rock and Mine.

workflow SONAR data -> Data processing -> train,test and split Data. when new data will be given to trained logistic Regression model it will predict object is Rock or Mine.

Logistic Regression Model is used because it works well for binary classification problem.

Supervised learning Algorithm